+++ packages/Flow/src/package-test.ts
@@ -12,6 +12,7 @@ import './tests/order-edge-tests';
 import './tests/minimap-tests';
 import './tests/creation-script-emit-tests';
 import './tests/function-browser-tests';
+import './tests/inspect-tests';
 
 export let _package = new DG.Package();
 export { tests };